requirements.txt
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mimerender==0.6.0
python-dateutil==2.5.3
python-mimeparse==1.5.2
six==1.10.0
SQLAlchemy==1.0.12
visitor==0.1.2
Werkzeug==0.11.9
